requirements.txt
streamlit==1.38.0
PuLP==2.9.0
PyYAML==6.0.2
geopandas[all]==1.0.1
streamlit-folium==0.23.1
